Check ELB expired SSL certificates
Lambda function to interrogate ELB listeners and determine if a SSL certificate is going to expire. If the certificate already expired or is about to, send an email to the emailContact tag listed in the ELB.

This lambda function was written in Python.

Note: Account ID's, role names, and other sensitive data has been hidden for publication manners.